+++ vector/src/simd_extensions/mod.rs
@@ -201,22 +201,22 @@ macro_rules! simd_generic_impl {
         }
     }
 }
-#[cfg(any(feature = "doc", feature="avx"))]
+#[cfg(any(feature = "doc", feature="use_avx"))]
 pub mod avx;
 
-#[cfg(any(feature = "doc", feature="avx"))]
+#[cfg(any(feature = "doc", feature="use_avx"))]
 pub use self::avx::{Reg32, Reg64};
 
-#[cfg(any(feature = "doc", all(feature = "sse", not(feature = "avx"))))]
+#[cfg(any(feature = "doc", all(feature = "use_sse", not(feature = "use_avx"))))]
 pub mod sse;
 
-#[cfg(any(feature = "doc", all(feature = "sse", not(feature = "avx"))))]
+#[cfg(any(feature = "doc", all(feature = "use_sse", not(feature = "use_avx"))))]
 pub use self::sse::{Reg32, Reg64};
 
-#[cfg(any(feature = "doc", not(any(feature = "avx", feature="sse"))))]
+#[cfg(any(feature = "doc", not(any(feature = "use_avx", feature="use_sse"))))]
 pub mod fallback;
 
-#[cfg(any(feature = "doc", not(any(feature = "avx", feature="sse"))))]
+#[cfg(any(feature = "doc", not(any(feature = "use_avx", feature="use_sse"))))]
 pub use self::fallback::{Reg32, Reg64};
 
 simd_generic_impl!(f32, Reg32);